hb/requirements.txt
openpyxl==3.1.5
PyYAML==6.0.2
fpdf2==2.7.9
cryptography==43.0.1
pandas==2.2.3
numpy==2.1.3
fastapi==0.115.7
uvicorn==0.30.6
python-multipart==0.0.9
paho-mqtt==2.1.0
